deps/v8/test/webkit/fast/js/string-capitalization-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This test checks that toLowerCase and toUpperCase handle certain non-trivial cases correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL String("A𐐀").toLowerCase() should be a𐐨. Was a𐐀.
FAIL String("a𐐨").toUpperCase() should be A𐐀. Was A𐐨.
PASS String("ΚΟΣΜΟΣ ΚΟΣΜΟΣ").toLowerCase() is "κοσμος κοσμος"
PASS String("ß").toUpperCase() is "SS"
PASS String("ŉ").toUpperCase() is "ʼN"
PASS String("ǰ").toUpperCase() is "J̌"
PASS String("ﬃ").toUpperCase() is "FFI"
PASS String("FFI").toLowerCase() is "ffi"
PASS String("Ĳ").toLowerCase() is "ĳ"
PASS expected[String.fromCharCode(0xA78D).toLowerCase()] is true
PASS expected[String.fromCharCode(0x0265).toUpperCase()] is true
PASS expected[String.fromCharCode(0x10C7).toLowerCase()] is true
PASS expected[String.fromCharCode(0x2D27).toUpperCase()] is true
PASS expected[String.fromCharCode(0x2D2D).toLowerCase()] is true
PASS expected[String.fromCharCode(0x10CD).toUpperCase()] is true
PASS expected[String.fromCharCode(0x2CF2).toLowerCase()] is true
PASS expected[String.fromCharCode(0x2CF3).toUpperCase()] is true
PASS expected[String.fromCharCode(0xA792).toLowerCase()] is true
PASS expected[String.fromCharCode(0xA793).toUpperCase()] is true
PASS expected[String.fromCharCode(0xA7AA).toLowerCase()] is true
PASS expected[String.fromCharCode(0x0266).toUpperCase()] is true
PASS successfullyParsed is true

TEST COMPLETE
